requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==6.5.0
groq==0.4.0
langchain==0.0.335
langchain-groq==0.0.1
upstash-vector==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.1
django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
bcrypt==4.0.1
